// src/app/products/state/models/product.actiontypes.ts
export enum ProductActionTypes{
    ToggleProductCode = '[Product] Toggle Product',
    SetCurrentProduct = '[Product] Set Current Product',
    ClearCurrentProduct = '[Product] Clear Current Product',
    IntializeCurrentProduct = '[Product] Intialize Current Product',
    Load = '[Product] Load',
    LoadSuccess = '[Product] Load Success',
    LoadFail = '[Product] Load Fail',
    UpdateProduct = '[Product] Update',
    UpdateSuccess = '[Product] Update Success',
    UpdateFail = '[Product] Update Fail'

}